requirements.txt
flask==2.0.1
numpy==1.21.0
pandas==1.3.0
pywebview==3.7.2
pyinstaller==5.13.0
gunicorn==20.1.0
werkzeug==2.0.1
itsdangerous==2.0.1
click==8.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1